requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
bookwyrm @ git+https://github.com/deepfates/bookwyrm.git@942530422faaef877bb779c2302f0f21a9570dc9
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cog==0.9.9
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.16
fastapi==0.98.0
fastjsonschema==2.19.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mistune==3.0.2
multidict==6.0.5
mypy==1.10.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nltk==3.8.1
numpy==1.26.4
openai==1.31.0
packaging==24.0
pandocfilters==1.5.1
platformdirs==4.2.2
pydantic==1.10.15
pydantic_core==2.18.4
Pygments==2.18.0
PyPDF2==3.0.1
pyperclip==1.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
replicate==0.26.0
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
ruff==0.4.7
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.27.0
structlog==24.2.0
tenacity==8.3.0
tiktoken==0.7.0
tinycss2==1.3.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
types-beautifulsoup4==4.12.0.20240511
types-html5lib==1.1.11.20240228
types-requests==2.32.0.20240602
types-tqdm==4.66.0.20240417
typing_extensions==4.12.1
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
webencodings==0.5.1
websockets==12.0
wget==3.2
yarl==1.9.4
youtube-transcript-api==0.6.2
